+++ miniscope_io/exceptions.py
@@ -13,3 +13,28 @@ class EndOfRecordingException(StopIteration):
     """
     Raised when :class:`.io.SDCard` is at the end of the available recording!
     """
+
+class StreamError(RuntimeError):
+    """
+    Base class for errors while streaming data
+    """
+
+class StreamReadError(StreamError):
+    """
+    Error while reading streaming data from a device
+    """
+
+class DeviceError(RuntimeError):
+    """
+    Base class for errors when communicating with or configuring devices
+    """
+
+class DeviceOpenError(DeviceError):
+    """
+    Error opening a connection to a device
+    """
+
+class DeviceConfigurationError(DeviceError):
+    """
+    Error while configuring a device
+    """