+++ docs/meta/changelog.md
@@ -1,5 +1,62 @@
 # Changelog
 
+## 0.1.9 - 24-06-21
+
+- https://github.com/Aharoni-Lab/miniscope-io/pull/26
+
+Testing:
+
+- [@t-sasatani](https://github.com/t-sasatani) - add end-to-end test for {class}`~miniscope_io.stream_daq.streamDaq`
+- Add a mock class for {class}`~miniscope_io.devices.opalkelly.okDev`
+- replace `tmpdir` fixture and `tempfile` module with `tmp_path`
+
+New:
+
+- [@t-sasatani](https://github.com/t-sasatani) - allow use of okDev on Windows
+- {meth}`~miniscope_io.stream_daq.StreamDaq.capture` can export video :)
+- More specific exceptions:
+  - {class}`~miniscope_io.exceptions.StreamError`
+  - {class}`~miniscope_io.exceptions.StreamReadError`
+  - {class}`~miniscope_io.exceptions.DeviceError`
+  - {class}`~miniscope_io.exceptions.DeviceOpenError`
+  - {class}`~miniscope_io.exceptions.DeviceConfigurationError`
+- {func}`~miniscope_io.utils.hash_video` - hash decoded video frames, rather than encoded video file
+
+
+Fixed:
+
+- Removed `print` statements in {class}`~miniscope_io.devices.opalkelly.okDev`
+- {meth}`~miniscope_io.stream_daq.StreamDaq.capture`
+  - Don't require `config`
+  - Replace logging with {func}`~miniscope_io.logging.init_logger`
+  - Use of {attr}`~miniscope_io.stream_daq.StreamDaq.terminate` to control inner loops
+
+
+Models:
+
+- added `fs` and `show_video` to {class}`~miniscope_io.models.stream.StreamDaqConfig`
+
+CI:
+
+- [@t-sasatani](https://github.com/t-sasatani) - restore windows and mac tests (oops)
+- caching dependency installs
+- not using pytest-emoji, it was always annoying
+
+## 0.1.8 - 24-06-16
+
+- https://github.com/Aharoni-Lab/miniscope-io/pull/21
+- https://github.com/Aharoni-Lab/miniscope-io/pull/15
+
+New features:
+
+- **Support for Various Image Formats**: `streamDaq` now supports multiple image formats, including different image sizes, frame rates (FPS), and bit-depths. These configurations can be provided via a YAML file. Examples of these configurations can be found in `miniscope_io.data.config`.
+- **Pydantic Model for Configuration**: Added a Pydantic model to validate the configuration of `streamDaq`.
+- **Bitstream Loader**: Added a bitstream loader to automatically configure the Opal Kelly FPGA when running `streamDaq`.
+- **Updated Command Line Script**: The command line script for running `streamDaq` has been updated. Use `streamDaq -c path/to/config/yaml/file.yml` to run the process with your YAML configuration file.
+- **Logger Module**: Added a logger module that can be configured using environmental variables or a `.env` file.
+
+Note: Version 0.1.7 was skipped accidentally and does not exist.
+
 ## 0.1.6 - 24-04-09
 
 - https://github.com/Aharoni-Lab/miniscope-io/pull/14